interactive editing, Code generation improvements, OpenAPISpec

+++ src/scripts/generate.js
@@ -1,42 +1,46 @@
 
 /**
- * node ./src/scripts/generate.js --modelName="bucket"
- * docker exec -ti express-api node ./src/scripts/generate.js --modelName="bucket"
-
+ * node ./src/scripts/generate.js --model="bucketCase"
+ * docker exec -ti express-api node ./src/scripts/generate.js --model="bucket"
  */
 require('dotenv').config();
 const argv = require('minimist')(process.argv.slice(2));
+const jsConvert = require('js-convert-case');
+const Inflector = require('inflector-js');
 const { v4: uuidv4 } = require('uuid');
 const Mustache = require('mustache');
 const moment = require('moment');
 const path = require('path');
 const fs = require('fs');
 
-if (!argv['modelName']) throw Error('You must provide --modelName argument');
+if (!argv['model']) throw Error('You must provide --model argument');
+if (/^\d/.test(argv['model'])) throw Error('--model cannot start with a number');
 
 (async function Main() {
-    const ucFirst = (string) => (string.charAt(0).toUpperCase().concat(string.slice(1)));
+    const modelName = argv['model'];
+    const isDryRun = !!argv['d'] || false;
 
     const params = {
-        modelname: argv['modelName'].toLowerCase(),
-        modelName: argv['modelName'],
-        ModelName: ucFirst(argv['modelName']),
-        MODELNAME: argv['modelName'].toUpperCase(),
+        modelname: jsConvert.toLowerCase(modelName),
+        modelName: jsConvert.toCamelCase(modelName),
+        ModelName: jsConvert.toPascalCase(modelName),
+        MODELNAME: jsConvert.toUpperCase(modelName),
 
-        modelnames: argv['modelName'].toLowerCase().concat('s'),
-        modelNames: argv['modelName'].concat('s'),
-        ModelNames: ucFirst(argv['modelName']).concat('s'),
+        modelnames: Inflector.pluralize(jsConvert.toLowerCase(modelName)),
+        modelNames: Inflector.pluralize(jsConvert.toCamelCase(modelName)),
+        ModelNames: Inflector.pluralize(jsConvert.toPascalCase(modelName)),
         UUID: uuidv4(),
     };
 
     if (argv['v']) console.log(params);
+    if (isDryRun) console.log(`\n\n⚠️  Dry run. Will not write any files.\n\n`);
 
 
     ////////////////////////////////////////////////
     // Model
     const pathModel = path.resolve(`./src/models/${params.ModelName}.js`);
     if (!argv['force'] && fs.existsSync(pathModel)) throw new Error(`File already exists at ${pathModel}`);
-    fs.writeFileSync(pathModel, Mustache.render(fs.readFileSync(path.resolve('./src/scripts/generator/Model.js'), 'utf8'), params));
+    if (!isDryRun) fs.writeFileSync(pathModel, Mustache.render(fs.readFileSync(path.resolve('./src/scripts/generator/Model.js'), 'utf8'), params));
     console.log(`Created: ${pathModel}`);
 
 
@@ -69,7 +73,7 @@ if (!argv['modelName']) throw Error('You must provide --modelName argument');
             ]).join('\n');
             lines.splice(lastLineWithClosingBracket - 1, 1, newLines);
         }
-        fs.writeFileSync(modelsIndex, lines.join('\n'));
+        if (!isDryRun) fs.writeFileSync(modelsIndex, lines.join('\n'));
     }
 
 
@@ -77,7 +81,7 @@ if (!argv['modelName']) throw Error('You must provide --modelName argument');
     // Route
     const pathRoute = path.resolve(`./src/routes/${params.ModelNames}.js`);
     if (!argv['force'] && fs.existsSync(pathRoute)) throw new Error(`File already exists at ${pathRoute}`);
-    fs.writeFileSync(pathRoute, Mustache.render(fs.readFileSync(path.resolve('./src/scripts/generator/Route.js'), 'utf8'), params));
+    if (!isDryRun) fs.writeFileSync(pathRoute, Mustache.render(fs.readFileSync(path.resolve('./src/scripts/generator/Route.js'), 'utf8'), params));
     console.log(`Created: ${pathRoute}`);
 
     const mainIndex = path.resolve(`./src/index.js`);
@@ -95,7 +99,7 @@ if (!argv['modelName']) throw Error('You must provide --modelName argument');
         if (lastLineWithRouteRequirePlusOne) {
             const routeRequireLine = `app.use('/api/v1/', require('./routes/${params.ModelNames}'));\n`;
             mainIndexLines.splice(lastLineWithRouteRequirePlusOne - 1, 1, routeRequireLine);
-            fs.writeFileSync(mainIndex, mainIndexLines.join('\n'));
+            if (!isDryRun) fs.writeFileSync(mainIndex, mainIndexLines.join('\n'));
         }
     }
 
@@ -104,7 +108,7 @@ if (!argv['modelName']) throw Error('You must provide --modelName argument');
     // Migration
     const pathMigration = path.resolve(`./src/database/migrations/${moment().format('YYYYMMDDHHmmss')}-create-${params.ModelNames}.js`);
     if (!argv['force'] && fs.existsSync(pathMigration)) throw new Error(`File already exists at ${pathMigration}`);
-    fs.writeFileSync(pathMigration, Mustache.render(fs.readFileSync(path.resolve('./src/scripts/generator/Migration.js'), 'utf8'), params));
+    if (!isDryRun) fs.writeFileSync(pathMigration, Mustache.render(fs.readFileSync(path.resolve('./src/scripts/generator/Migration.js'), 'utf8'), params));
     console.log(`Created: ${pathMigration}`);
 
 
@@ -112,7 +116,7 @@ if (!argv['modelName']) throw Error('You must provide --modelName argument');
     // Seeder
     const pathSeeder = path.resolve(`./src/database/seeders/${moment().format('YYYYMMDDHHmmss')}-${params.ModelNames}.js`);
     if (!argv['force'] && fs.existsSync(pathSeeder)) throw new Error(`File already exists at ${pathSeeder}`);
-    fs.writeFileSync(pathSeeder, Mustache.render(fs.readFileSync(path.resolve('./src/scripts/generator/Seeder.js'), 'utf8'), params));
+    if (!isDryRun) fs.writeFileSync(pathSeeder, Mustache.render(fs.readFileSync(path.resolve('./src/scripts/generator/Seeder.js'), 'utf8'), params));
     console.log(`Created: ${pathSeeder}`);
 
 
@@ -120,7 +124,7 @@ if (!argv['modelName']) throw Error('You must provide --modelName argument');
     // Test
     const pathTest = path.resolve(`./tests/${params.ModelName}.js`);
     if (!argv['force'] && fs.existsSync(pathTest)) throw new Error(`File already exists at ${pathTest}`);
-    fs.writeFileSync(pathTest, Mustache.render(fs.readFileSync(path.resolve('./src/scripts/generator/Test.js'), 'utf8'), params));
+    if (!isDryRun) fs.writeFileSync(pathTest, Mustache.render(fs.readFileSync(path.resolve('./src/scripts/generator/Test.js'), 'utf8'), params));
     console.log(`Created: ${pathTest}`);
 
 
@@ -130,7 +134,7 @@ if (!argv['modelName']) throw Error('You must provide --modelName argument');
     const requestContent = fs.readFileSync(pathRequests, 'utf8');
     const newRequest = Mustache.render(fs.readFileSync(path.resolve('./src/scripts/generator/requests.http'), 'utf8'), params);
     if (!requestContent.includes(`### ${params.ModelName}`)) {
-        fs.writeFileSync(pathRequests, requestContent + '\n' + newRequest);
+        if (!isDryRun) fs.writeFileSync(pathRequests, requestContent + '\n' + newRequest);
         console.log(`Updated: ${pathRequests}`);
     }
 })();
